Moved the chain rule to a method in the returned object
This does not work because of a circular import error, reverting
